main: Improved _monitor_repo logic
The If-None-Match header is added to the request with the value of the last received ETag header. If the resource has not been updated since the last request, the API returns a 304 Not Modified response, and the method sleeps for 60 seconds before making the next request. If the resource has been updated, the new data is processed and the new ETag header is saved for use in the next request.

Added a retries variable to keep track of the number of retries. If there is a aiohttp.ClientError exception, we retry the operation up to five times with increasing wait times (60 seconds for the first retry, 120 seconds for the second retry, etc.). If the operation still fails after five retries, we raise an exception to stop the loop and notify the user. For other types of exceptions, we raise an exception immediately to stop the loop and notify the user.